+++ src/main/java/org/kde/kdeconnect/plugins/mousereceiver/MouseReceiverPlugin.java
@@ -57,10 +57,10 @@ public class MouseReceiverPlugin extends Plugin {
             return false; // This packet will be handled by the remotekeyboard instead, silently ignore
         }
 
-        double dx = np.getDouble("dx", 0);
-        double dy = np.getDouble("dy", 0);
-        double x = np.getDouble("x", 0);
-        double y = np.getDouble("y", 0);
+        int dx = np.getInt("dx", 0);
+        int dy = np.getInt("dy", 0);
+        int x = np.getInt("x", 0);
+        int y = np.getInt("y", 0);
 
         boolean isSingleClick = np.getBoolean("singleclick", false);
         boolean isDoubleClick = np.getBoolean("doubleclick", false);
@@ -112,7 +112,7 @@ public class MouseReceiverPlugin extends Plugin {
             // Mouse Move
             if (dx != 0 || dy != 0) {
                 // Log.i("MouseReceiverPlugin", "move Mouse dx: " + dx + " dy: " + dy);
-                return MouseReceiverService.move(dx, dy);
+                return MouseReceiverService.move((int)dx, (int)dy);
             } else if (x != 0 || y != 0) {
                 return MouseReceiverService.setPos(x, y);
             } else {
